Refs #136. Moved vanadium processing UI from reduced data view to peak processing UI.
